Refactored to rename no-curly-component-invocation
