+++ lib/helpers/curly-component-invocation.js
