+++ lib/rules/index.js
@@ -17,7 +17,7 @@ module.exports = {
   'no-action-modifiers': require('./lint-no-action-modifiers'),
   'no-attrs-in-components': require('./lint-no-attrs-in-components'),
   'no-bare-strings': require('./lint-no-bare-strings'),
-  'no-classic-invocation-component': require('./lint-no-classic-invocation-component'),
+  'no-curly-component-invocation': require('./lint-no-curly-component-invocation'),
   'no-debugger': require('./lint-no-debugger'),
   'no-duplicate-attributes': require('./lint-duplicate-attributes'),
   'no-element-event-actions': require('./lint-no-element-event-actions'),
