+++ lib/rules/lint-no-curly-component-invocation.js
@@ -3,7 +3,7 @@ const {
   transformTagName,
   isComponentTagName,
   isNestedComponentTagName,
-} = require('../helpers/classic-invocation-component');
+} = require('../helpers/curly-component-invocation');
 
 module.exports = class NoComponentClassicInvocation extends Rule {
   logNode({ message, node }) {
@@ -28,7 +28,7 @@ module.exports = class NoComponentClassicInvocation extends Rule {
 
   generateError(name) {
     let angleBracketName = transformTagName(name);
-    return `You are using the component {{${name}}} with classic invocation syntax. You should use <${angleBracketName}> instead. If it is actually a helper you must manually add it to the 'no-classic-invocation-component' rule configuration, e.g. \`'no-classic-invocation-component': { allow: ['${name}'] }\`.`;
+    return `You are using the component {{${name}}} with curly component syntax. You should use <${angleBracketName}> instead. If it is actually a helper you must manually add it to the 'no-curly-component-invocation' rule configuration, e.g. \`'no-curly-component-invocation': { allow: ['${name}'] }\`.`;
   }
 
   visitor() {
